Refactored linformer into a single file
